requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
yfinance>=0.2.20
PyYAML>=6.0

# Development dependencies
pytest>=7.0.0